deploy/Gemini-Summary/requirements.txt
pika==1.3.1
mysql-connector-python==8.4.0
google-genai